airplane_system/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
reportlab==4.2.2
qrcode[pil]==7.4.2
Pillow==10.3.0
stripe==8.10.0
